requirements.txt
asgiref==3.5.2
black==22.6.0
build==0.8.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pep517==0.13.0
pgeocode==0.3.0
pip-tools==6.8.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
drf-yasg==1.21.3
